chore(package.json): add package.json file for graphql-server-example project
A new package.json file is added for the graphql-server-example project. This file includes the necessary metadata such as name, version, description, main file, scripts, keywords, author, license, and dependencies. The dependencies include "@apollo/server" version 4.9.5 and "graphql" version 16.8.1. The type is set to "module" to enable ES modules in the project.